feat: adds GetPermissionForRoleAPI + tests (#432)
* initial commit

* query fixs, adds storage tests

* adds userRoles tests

* adds removeUserRoleAPI

* adds removeUserRoleAPI

* updates RemoveUserRole api and function to now return a boolean on whether the user had the role

* adds serialVersion field for API

* adds test to check removing a role from a user, where the user did not have the role

* adds more core tests

* fixs

* adds FOR UPDATE to doesRoleExist_Transaction

* udpates to test

* fixs

* fixs

* test fixs

* adds test and query fix

* adds remaining storage test

* adds core function test

* temp commit

* fixs

* test fixs

* test fixs

* typo fix

* feat: adds GetPermissionsForRoleAPI + tests

* fixs

* removes unnecessary print statments

* adds tests and some fixs

+++ src/test/java/io/supertokens/test/userRoles/api/GetPermissionsForRoleAPITest.java
@@ -18,7 +18,6 @@ package io.supertokens.test.userRoles.api;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
 import io.supertokens.ProcessState;
 import io.supertokens.pluginInterface.STORAGE_TYPE;
 import io.supertokens.storageLayer.StorageLayer;
@@ -33,13 +32,11 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TestRule;
 
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.HashMap;
 
 import static org.junit.Assert.*;
 
-public class GetUserRolesAPITest {
+public class GetPermissionsForRoleAPITest {
     @Rule
     public TestRule watchman = Utils.getOnFailure();
 
@@ -68,27 +65,42 @@ public class GetUserRolesAPITest {
             // no request params
             try {
                 HttpRequestForTesting.sendGETRequest(process.getProcess(), "",
-                        "http://localhost:3567/recipe/user/roles", new HashMap<>(), 1000, 1000, null,
+                        "http://localhost:3567/recipe/role/permissions", new HashMap<>(), 1000, 1000, null,
                         Utils.getCdiVersion2_14ForTests(), "userroles");
                 throw new Exception("should not come here");
             } catch (HttpResponseException e) {
                 assertTrue(e.statusCode == 400 && e.getMessage().equals(
-                        "Http error. Status Code: 400. Message:" + " Field name 'userId' is missing in GET request"));
+                        "Http error. Status Code: 400. Message:" + " Field name 'role' is missing in GET request"));
             }
         }
 
         {
-            // userId is missing
+            // role is missing
             HashMap<String, String> QUERY_PARAM = new HashMap<>();
             QUERY_PARAM.put("invalid", "invalid");
             try {
                 HttpRequestForTesting.sendGETRequest(process.getProcess(), "",
-                        "http://localhost:3567/recipe/user/roles", QUERY_PARAM, 1000, 1000, null,
+                        "http://localhost:3567/recipe/role/permissions", QUERY_PARAM, 1000, 1000, null,
                         Utils.getCdiVersion2_14ForTests(), "userroles");
                 throw new Exception("should not come here");
             } catch (HttpResponseException e) {
                 assertTrue(e.statusCode == 400 && e.getMessage().equals(
-                        "Http error. Status Code: 400. Message:" + " Field name 'userId' is missing in GET request"));
+                        "Http error. Status Code: 400. Message:" + " Field name 'role' is missing in GET request"));
+            }
+        }
+
+        {
+            // role is an empty string
+            HashMap<String, String> QUERY_PARAM = new HashMap<>();
+            QUERY_PARAM.put("role", " ");
+            try {
+                HttpRequestForTesting.sendGETRequest(process.getProcess(), "",
+                        "http://localhost:3567/recipe/role/permissions", QUERY_PARAM, 1000, 1000, null,
+                        Utils.getCdiVersion2_14ForTests(), "userroles");
+                throw new Exception("should not come here");
+            } catch (HttpResponseException e) {
+                assertTrue(e.statusCode == 400 && e.getMessage().equals(
+                        "Http error. Status Code: 400. Message:" + " Field name 'role' cannot be an empty String"));
             }
         }
 
@@ -98,7 +110,7 @@ public class GetUserRolesAPITest {
     }
 
     @Test
-    public void testGettingRolesForAUser() throws Exception {
+    public void testRetrievingThePermissionsForARole() throws Exception {
         String[] args = { "../" };
 
         TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
@@ -108,46 +120,51 @@ public class GetUserRolesAPITest {
             return;
         }
 
-        // create and add multiple roles to a user, get roles for user
-        {
-            String[] roles = new String[] { "role1", "role2", "role3" };
-            String userId = "userId";
+        // create a role with permissions
+        String role = "role";
+        String[] permissions = new String[] { "permission1", "permission2", "permission3" };
+        UserRoles.createNewRoleOrModifyItsPermissions(process.main, role, permissions);
 
-            for (String role : roles) {
-                // create role
-                UserRoles.createNewRoleOrModifyItsPermissions(process.main, role, null);
-                // add role to User
-                UserRoles.addRoleToUser(process.main, userId, role);
-            }
+        HashMap<String, String> QUERY_PARAM = new HashMap<>();
+        QUERY_PARAM.put("role", role);
 
-            HashMap<String, String> QUERY_PARAMS = new HashMap<>();
-            QUERY_PARAMS.put("userId", userId);
-            JsonObject response = HttpRequestForTesting.sendGETRequest(process.getProcess(), "",
-                    "http://localhost:3567/recipe/user/roles", QUERY_PARAMS, 1000, 1000, null,
-                    Utils.getCdiVersion2_14ForTests(), "userroles");
+        JsonObject response = HttpRequestForTesting.sendGETRequest(process.getProcess(), "",
+                "http://localhost:3567/recipe/role/permissions", QUERY_PARAM, 1000, 1000, null,
+                Utils.getCdiVersion2_14ForTests(), "userroles");
 
-            assertEquals(2, response.entrySet().size());
-            assertEquals("OK", response.get("status").getAsString());
+        assertEquals(2, response.entrySet().size());
+        assertEquals("OK", response.get("status").getAsString());
 
-            JsonArray userRolesArr = response.getAsJsonArray("roles");
-            String[] userRoles = Utils.parseJsonArrayToStringArray(userRolesArr);
-            Utils.checkThatArraysAreEqual(roles, userRoles);
+        JsonArray arr = response.get("permissions").getAsJsonArray();
+        String[] retrievedPermissions = Utils.parseJsonArrayToStringArray(arr);
+
+        // check that retrieved permissions are the same as the ones set
+        Utils.checkThatArraysAreEqual(permissions, retrievedPermissions);
+
+        process.kill();
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+    }
+
+    @Test
+    public void testRetrievingThePermissionsForAnUnknownRole() throws Exception {
+        String[] args = { "../" };
+
+        TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
+
+        if (StorageLayer.getStorage(process.getProcess()).getType() != STORAGE_TYPE.SQL) {
+            return;
         }
 
-        // get roles for a user who has no roles
-        {
-            String userId2 = "userId2";
-            HashMap<String, String> QUERY_PARAMS = new HashMap<>();
-            QUERY_PARAMS.put("userId", userId2);
+        HashMap<String, String> QUERY_PARAM = new HashMap<>();
+        QUERY_PARAM.put("role", "unknownRole");
 
-            JsonObject response = HttpRequestForTesting.sendGETRequest(process.getProcess(), "",
-                    "http://localhost:3567/recipe/user/roles", QUERY_PARAMS, 1000, 1000, null,
-                    Utils.getCdiVersion2_14ForTests(), "userroles");
+        JsonObject response = HttpRequestForTesting.sendGETRequest(process.getProcess(), "",
+                "http://localhost:3567/recipe/role/permissions", QUERY_PARAM, 1000, 1000, null,
+                Utils.getCdiVersion2_14ForTests(), "userroles");
 
-            assertEquals(2, response.entrySet().size());
-            assertEquals("OK", response.get("status").getAsString());
-            assertEquals(0, response.getAsJsonArray("roles").size());
-        }
+        assertEquals(1, response.entrySet().size());
+        assertEquals("UNKNOWN_ROLE_ERROR", response.get("status").getAsString());
 
         process.kill();
         assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
